refactor: Separate the code in src directory
