+++ config/network-parameters.js
@@ -1,9 +1,8 @@
 const DIFFICULTY = 3;
 const CREATION_TIME = 30000;
 const BLOCK_SIZE = 3;
-const TEST_KEY = '0c9cf4484e0925c2b178fbdeea967b303b065ab693d54941307f9c96f359d24b';
 const MINING_REWARD = 1;
 
 module.exports = {
-    DIFFICULTY, CREATION_TIME, BLOCK_SIZE, TEST_KEY, MINING_REWARD,
+    DIFFICULTY, CREATION_TIME, BLOCK_SIZE, MINING_REWARD
 };
